refactor: migrate commands to repositories, simplify DTOs, and apply Pint code style fixes

+++ resources/views/livewire/soapnuts-balance.blade.php
@@ -1,7 +1,7 @@
 <div>
     @auth
         <span class="text-gold-400 font-semibold">
-        {{ number_format((int) auth()->user()->soapnuts) }} 🌰
+        {{ number_format(auth()->user()->soapnuts) }} 🌰
     </span>
     @endauth
 </div>
